Colocate diagnostics classes, drop DataOp, rename `mode` to `type`
Each diagnostics class now lives next to the class that publishes it - `StoreDiagnostics` beside
`Store`, `ViewDiagnostics` beside `View`, `GridDiagnostics` beside `GridModel` - each carrying its
own op interface and type union. Drops the shared `data/Diagnostics.ts`.

Removes the `DataOp` base interface. It only ever contributed three fields and forced readers
through a level of indirection to find them; each op interface now declares them outright. The
stability contract it documented moves onto each diagnostics class, where it is visible at the
point of use.

Renames the `mode` field to `type`, matching the vocabulary used elsewhere in the data package.
