+++ data/index.ts
@@ -5,9 +5,9 @@
  * Copyright © 2026 Extremely Heavy Industries Inc.
  */
 
-export * from './Diagnostics';
 export * from './Field';
 export * from './Store';
+export * from './StoreDiagnostics';
 export * from './StoreRecord';
 export * from './RecordAction';
 export * from './StoreSelectionModel';
@@ -38,6 +38,7 @@ export * from './cube/Cube';
 export * from './cube/CubeField';
 export * from './cube/Query';
 export * from './cube/View';
+export * from './cube/ViewDiagnostics';
 export * from './cube/ViewRowData';
 
 export * from './validation/constraints';
